moved lang specific ES tests into seperate test file so that indexes could be reset, corrected insert logic
